Implement rational tree unification algorithm
In this commit tests are broken.

This commit consists of three main parts:
1. Change of the substitution interpretation (but not representation)
2. Change of the substitution application process to support recursion
   and mu-binders introduction
3. Change of the unification algorithm to allow rational terms in
   substitution (i.e. recursive equation solution)

=== Subsitution interpretation ===

Term head is a term constructor with only logic variables in place of
arguments.

We interpret substitution as a system of possibly recursive equations
on terms in special form. Left hand sides of that system must
be distinct logic variables (as in a regular substitution) and right
hand sides must be either a logic variable or a term head.

Invariant: any chain of equations with logic variable in left hand sides
must not be cyclic. In other words, if we have: x1 = x2, x2 = x3, ...,
x_{n - 1} = x_n, all x_i must be distinct. The unification algorithm
guarantees it.

We have two main operations on equation systems:
- Find (named `walk` in source code) fetches the last variable in a
  chain of the given variable and a bound term head if it's exist
- Union populates the equation system with new equation to equalize two
  given logic variables and returns a pair of term heads if the both of
  them already bound with them. It does nothing if the given variables
  are already present in the same chain.

=== Substitution application ===

We define substitution application as a homomorphic extension of
substitution image, i.e. we just map the given term with the image
function.

We generate an image of the given logic variable using auxiliary data
structure 'visited' to prevent infinite recursion and to determine when
we need to insert a mu-binder. We just unfold every occurred variable
only once and if the image of variable contains the variable itself,
insert a mu-binder outside of them image.

=== Unification algorithm ===

Given two terms, we just fold them as a regular, but with special
handling of variable-variable and variable-term cases. In case of
variable-variable we just union them first, then unify returned pair of
terms. It guarantees that we will not unify one pair of variables twice.
In case of variable-term we check have the variable a binding, and in
this case just unify the binding, otherwise inject the given term.
If the binding exists, we guarantee that it just a term head so we must
face a leaf (variable-variable or variable-constant) in this recursion.

Injection is a corner case of the unification algorithm that just
converts complex term into a serie of term heads with fresh variables to
save an invariant.

Signed-off-by: Eridan Domoratskiy <eridan200@mail.ru>