Fix 'State.reify' to apply subst on variable constraints
In the previous implementation there wasn't explicit substitution
application, but for some reasons terms in a disequality constraint
storage was already with applied substitution in many cases (but some
tests show that it isn't that always) and seems like it was acceptable.
But in the current implementation a constraint storage doesn't include
full terms frequently so substitution application is being significant.

Signed-off-by: Eridan Domoratskiy <eridan200@mail.ru>